pyproject.toml
[tool.poetry]
name = "agentlens"
version = "0.1.0"
description = "轻量级 Agent 可观测平台"
authors = ["OpenClaw Team <team@openclaw.ai>"]
readme = "README.md"
packages = [{include = "agentlens", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0"
httpx = "^0.27"
structlog = "^24.1"
python-dotenv = "^1.0"
click = "^8.1"
rich = "^13.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
black = "^24.0"
ruff = "^0.3"
mypy = "^1.8"

[tool.poetry.scripts]
agentlens = "agentlens.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
